Update ruby version and add db cleaner

+++ lib/tasks/missql.rake
@@ -1,7 +1,7 @@
 namespace :missql do
 	desc "Drop and recreate the development level db"
 	task :reset => :environment do
-    User.destroy_all
+    User.destroy_all if ActiveRecord::Base.connection.table_exists? 'users'
 		Rake::Task["db:drop"].invoke
 		Rake::Task["db:create"].invoke
 		Rake::Task["db:migrate"].invoke
